membenarkan baim

+++ app/Http/Controllers/Api/AuthController.php
@@ -37,7 +37,9 @@ class AuthController extends Controller
         return response()->json([
             'message' => "Login berhasil, Anda login sebagai {$user->role->nama_role}",
             'token' => $token,
-            'dashboard_url' => $dashboardUrl
+            'dashboard_url' => $dashboardUrl,
+            'role' => $role
+
         ]);
     }
 
@@ -117,6 +119,10 @@ class AuthController extends Controller
     }
     public function logout(Request $request)
     {
+        if (!$request->user()) {
+            return response()->json(['message' => 'User tidak terautentikasi'], 401);
+        }
+
         $request->user()->tokens()->delete(); // Hapus semua token user ini
         return response()->json([
             'message' => 'Logout berhasil.'
